+++ src/index.js
@@ -6,6 +6,7 @@ export { default as Button } from './components/Elements/Button';
 export { default as ButtonGroup } from './components/Elements/ButtonGroup';
 export { default as Icon } from './components/Elements/Icon';
 export { default as Pill } from './components/Elements/Pill';
+export { default as Placeholder } from './components/Elements/Placeholder';
 export { default as Spinner } from './components/Elements/Spinner';
 export { default as Tab } from './components/Elements/Tab';